requirements.txt
alembic==1.13.1
annotated-types==0.6.0
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
decouple==0.0.7
discord-webhook==1.3.1
easyocr==1.7.1
filelock==3.13.1
fsspec==2024.2.0
greenlet==3.0.3
idna==3.6
imageio==2.34.0
Jinja2==3.1.3
keyboard==0.13.5
lazy_loader==0.3
loguru==0.7.2
Mako==1.3.2
MarkupSafe==2.1.5
MouseInfo==0.1.3
mpmath==1.3.0
mss==9.0.1
networkx==3.2.1
ninja==1.11.1.1
numpy==1.26.4
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
packaging==24.0
pandas==2.2.1
pillow==10.2.0
psutil==5.9.8
PyAutoGUI==0.9.54
pyclipper==1.3.0.post5
pydantic==2.6.3
pydantic_core==2.16.3
PyDirectInput==1.0.4
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pynput==1.7.6
pyperclip==1.8.2
PyQt6==6.4.2
PyQt6-Qt6==6.6.2
PyQt6-sip==13.6.0
PyRect==0.2.0
PyScreeze==0.1.30
python-bidi==0.4.2
python-dateutil==2.9.0.post0
python-decouple==3.8
pytweening==1.2.0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
scikit-image==0.22.0
scipy==1.12.0
shapely==2.0.3
six==1.16.0
SQLAlchemy==2.0.28
sqlmodel==0.0.16
sympy==1.12
tifffile==2024.2.12
torch==2.2.1
torchvision==0.17.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
win32-setctime==1.1.0
